+++ microservices/scraper-australia/.dockerignore
@@ -1,4 +1,4 @@
 *
 !requirements.txt
 !app
-!myapp.py
+!run.py
