+++ microservices/scraper-australia/Dockerfile
@@ -4,4 +4,4 @@ COPY requirements.txt .
 RUN pip install -r requirements.txt
 COPY . .
 ENV GUNICORN_CMD_ARGS="--bind=0.0.0.0"
-CMD ["gunicorn", "app"]
+CMD ["gunicorn", "run:app", "--timeout", "5000000000000", "--log-level=DEBUG"]
