+++ microservices/scraper-australia/app/__init__.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+app = Flask(__name__)
+from app import routes
