+++ microservices/scraper-australia/app/routes.py
@@ -1,7 +1,5 @@
-from flask import Flask
-from scraper import get_page_data
-
-app = Flask(__name__)
+from app import app
+from app.scraper import get_page_data
 
 
 @app.route("/routes")
@@ -12,7 +10,3 @@ def get_routes():
 @app.route("/data")
 def page_data():
     return str(get_page_data())
-
-
-if __name__ == "__main__":
-    app.run(debug=True)
