+++ microservices/scraper-australia/app/tests/example_scraper_test.py
@@ -1,0 +1,6 @@
+import pytest
+from app import scraper
+
+
+def test_basic():
+    assert scraper.get_page_data() is not None
