+++ microservices/scraper-australia/now.json
@@ -1,6 +1,7 @@
 {
     "public": true,
     "type": "docker",
+    "alias":"gg-scraper-example",
     "features": {
       "cloud": "v2"
     }
