+++ microservices/scraper-australia/requirements.txt
@@ -1,5 +1,4 @@
-requests==2.19.1
+Flask==0.11.1
+requests==2.9.1
 beautifulsoup4==4.6.3
-Flask==1.0.2
-microservices==0.33.0
-pytest==3.8.2
+gunicorn==19.9.0
